feat: add npm installation support, global language mode, and bug fixes
- Add npm package installation method
- Implement global language mode support
- Reorganize documentation structure to docs/ directory
- Fix various bugs and improve code quality
- Update project dependencies and configuration

+++ internal/ingestor/streaming.go
@@ -335,7 +335,7 @@ func (mm *MemoryMonitor) UpdateMemoryUsage() {
 
 	var m runtime.MemStats
 	runtime.ReadMemStats(&m)
-	mm.currentMemory = m.Alloc
+	mm.currentMemory = int64(m.Alloc)
 }
 
 // ShouldTriggerGC returns true if garbage collection should be triggered
